move onboarding into email account folder
