+++ apps/web/app/(app)/(redirects)/onboarding/page.tsx
@@ -1,0 +1,5 @@
+import { redirectToEmailAccountPath } from "@/utils/account";
+
+export default async function OnboardingPage() {
+  await redirectToEmailAccountPath("/onboarding");
+}
