+++ apps/web/app/(app)/[emailAccountId]/onboarding/OnboardingCategories.tsx
