+++ apps/web/app/(app)/[emailAccountId]/onboarding/StepIntro.tsx
@@ -4,9 +4,9 @@ import Image from "next/image";
 import { MailIcon } from "lucide-react";
 import { CardBasic } from "@/components/ui/card";
 import { PageHeading, TypographyP } from "@/components/Typography";
-import { IconCircle } from "@/app/(landing)/onboarding/IconCircle";
-import { OnboardingWrapper } from "@/app/(landing)/onboarding/OnboardingWrapper";
-import { ContinueButton } from "@/app/(landing)/onboarding/ContinueButton";
+import { IconCircle } from "@/app/(app)/[emailAccountId]/onboarding/IconCircle";
+import { OnboardingWrapper } from "@/app/(app)/[emailAccountId]/onboarding/OnboardingWrapper";
+import { ContinueButton } from "@/app/(app)/[emailAccountId]/onboarding/ContinueButton";
 
 export function StepIntro() {
   return (
